requirements.txt
pillow 
opencv-python 
gunicorn
numpy
torch 
transformers 
matplotlib
flask
keybert
sentencepiece
